fix main project reference to dedicated-worker
After reading https://www.typescriptlang.org/docs/handbook/project-references.html#what-is-a-project-reference,
especially this item:

    * Importing modules from a referenced project will instead load its output declaration file (.d.ts)

I finally got it.

It happens only if **referencing** project has **referenced** project
in its references.

It does not happen if all the projects are only referenced from the
toplevel project.

The usability could be improved.

Also, "composite" is supposed to be in "compilerOptions".

Also, verbose module resolution still reports as if dedicated-worker index.ts
still comesfrom the source dir, which can't be true.